Revert "Restructure PPAs according to new week structure"

+++ ppa/ppa_6.2/evaluation_file.sh
@@ -6,7 +6,7 @@ err(){
 }
 
 req=( "mktemp" "diff" "basename" "col" "find" "pushd" "popd" )
-executable="long.sh"
+executable="myCount.sh"
 for i in "${req[@]}"; do
   command -v "$i" > /dev/null 2>&1 || err "$i is not installed"
 done
@@ -26,8 +26,22 @@ cat >script.sh <<EOF
 rand_dir=\$(mktemp -d XXXXXX)
 pushd "\$rand_dir" > /dev/null || exit 1
 
-cat > data.txt
-bash "../\$(dirname "\${BASH_SOURCE[0]}")/$executable" 2>&1 < /dev/null
+cat > somefile.txt
+
+exec="../\$(dirname "\${BASH_SOURCE[0]}")/$executable"
+
+[[ -r "\$exec" ]] || exit 1
+
+if grep "wc|awk" "\$exec" &>/dev/null ; then
+  echo "Do not use wc or awk in your script"
+else
+  bash "\$exec" -l somefile.txt 2>&1 < /dev/null
+  bash "\$exec" -w somefile.txt 2>&1 < /dev/null
+  bash "\$exec" -n somefile.txt 2>&1 < /dev/null
+  bash "\$exec" -s say somefile.txt 2>&1 < /dev/null
+  bash "\$exec" -l -n somefile.txt 2>&1 < /dev/null
+  bash "\$exec" -l -s say -l -n somefile.txt 2>&1 < /dev/null
+fi
 
 popd > /dev/null || exit 1
 [[ -d "\$rand_dir" ]] && rm "\${rand_dir?}" -rf
