+++ ppa/ppa_7.2/evaluation_file.sh
@@ -6,7 +6,7 @@ err(){
 }
 
 req=( "mktemp" "diff" "basename" "col" "find" "pushd" "popd" )
-executable="investment.sh"
+executable="repeat.awk"
 for i in "${req[@]}"; do
   command -v "$i" > /dev/null 2>&1 || err "$i is not installed"
 done
@@ -26,29 +26,13 @@ cat >script.sh <<EOF
 rand_dir=\$(mktemp -d XXXXXX)
 pushd "\$rand_dir" > /dev/null || exit 1
 
-mkdir data || exit 1
+cat > myfile.txt
 
-read fnos
-for (( i=0; i<fnos; i++ )); do
-  read line
-  cut -d ":" -f 2- <<< "\$line" | tr '#' '\n' > ./data/\${line%%:*}
-done
+exec="../\$(dirname "\${BASH_SOURCE[0]}")/$executable"
 
-> map
-read mnos
-for (( i=0; i<mnos; i++ )); do
-  read line
-  echo "\$line"  >> map
-done
+[[ -r "\$exec" ]] || exit 1
 
-> result
-read rnos
-for (( i=0; i<rnos; i++ )); do
-  read line
-  echo "\$line"  >> result
-done
-
-bash "../\$(dirname "\${BASH_SOURCE[0]}")/$executable" 2>&1 < /dev/null
+awk -f "\$exec" myfile.txt 2>&1 < /dev/null
 
 popd > /dev/null || exit 1
 [[ -d "\$rand_dir" ]] && rm "\${rand_dir?}" -rf
